# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farah API Documentation",
        "description": "توثيق واجهات برمجة التطبيقات (APIs) لمنصة فرح لتطبيق Flutter",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://farahbackend-production.up.railway.app",
            "description": "سيرفر الإنتاج (Railway Production)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "السيرفر المحلي (Local Server)"
        }
    ],
    "paths": {
...
        "/api/provider/subscription-plans": {
            "get": {
                "tags": [
                    "Provider Subscriptions"
                ],
                "summary": "عرض باقات الاشتراك",
                "operationId": "cc6966cc9eb3db892ff0ec4d8d66cf2b",
                "responses": {
                    "200": {
                        "description": "قائمة الباقات"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/subscription": {
            "get": {
                "tags": [
                    "Provider Subscriptions"
                ],
                "summary": "عرض الاشتراك الحالي",
                "operationId": "caaad0261adb547b1338063ed7355152",
                "responses": {
                    "200": {
                        "description": "بيانات الاشتراك الحالي"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Provider Subscriptions"
                ],
                "summary": "تفعيل اشتراك مزود الخدمة",
                "operationId": "7581cccd69cd6b90c347f38d3d642010",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan"
                                ],
                                "properties": {
                                    "plan": {
                                        "type": "string",
                                        "example": "free-trial"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم تفعيل الاشتراك"
                    },
                    "409": {
                        "description": "تم استخدام الاشتراك المجاني مسبقًا"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/subscription/cancel": {
            "post": {
                "tags": [
                    "Provider Subscriptions"
                ],
                "summary": "إلغاء اشتراك مزود الخدمة",
                "operationId": "d8a5effcab98c28854405ade822db00b",
                "responses": {
                    "200": {
                        "description": "تم إلغاء الاشتراك"
                    },
                    "404": {
                        "description": "لا يوجد اشتراك فعال"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "أدخل التوكن بهذا الشكل: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}